+++ src/app/Page-Functionality/Navigations.js
@@ -1,48 +1,58 @@
 'use client'
 import { useContext } from "react";
-
 import { authContext } from "./Login/auth-context";
+import AddMenu from './AddMenu';
+import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 
 
 function Nav() {
   const { user, loading, logout } = useContext(authContext);
+  const handleClick = (action) => {
+    switch (action) {
+      case 'Graphs':
+        // Add logic 
+        break;
+      case 'support':
+        // Add logic 
+        break;
+      case 'account':
+        // Add logic 
+        break;
+      case 'calendar':
+        //Add logic
+        break;
+      default:
+        break;
+    }
+   };
 
-  return (
-    <header className="container max-w-2x1 px-6 mx-auto mt-3">
-      <div className="flex items-center justify-between">
-        {/* User information */}
-        {user && !loading && (
-          <div className="flex items-center gap-2">
-            {/* img */}
-            <div className="h-[40px] w-[40px] rounded-full overflow-hidden">
-              <img
-                className="object-cover w-full h-full"
-                src={user.photoURL}
-                alt={user.displayName}
-                referrerPolicy="no-referrer"
-              />
+   return (
+     <header className="container max-w-full px-20 mx-auto bg-custom-gray py-4">
+       <div className="flex items-center justify-between">
+         {/* Left side of the navigation */}
+         {user && !loading && (
+          <div className="flex items-center gap-4 ml-12">
+            <div>
+              <AddMenu />
             </div>
-
-            {/* name */}
-            <small>Hi, {user.displayName}!</small>
+            <span className="text-white ml-10 text-xl">Personal Finance Manager</span>
           </div>
         )}
-
-        {/* Right side of our navigation */}
+        {/* Right side of the navigation */}
         {user && !loading && (
           <nav className="flex items-center gap-4">
-            <div>
-            </div>
-            <div>
-              <button onClick={logout} className="btn btn-danger">
-                Sign out
-              </button>
-            </div>
+            <button onClick={handleClick('calendar')} className="btn mr-2">Calendar</button>
+            <button onClick={handleClick('graphs')} className="btn mr-2">Graphs</button>
+            <button onClick={handleClick('support')} className="btn mr-2">Support</button>
+            <button onClick={handleClick('account')} className="btn">Account</button>
+            <button onClick={logout} className="btn btn-danger mr-4">
+              Sign out
+            </button>
+            {/* Add more buttons as needed */}
           </nav>
         )}
       </div>
     </header>
   );
 }
-
 export default Nav;
